backend/requirements.txt
# ARIA MVP 2.0 — Backend Requirements

# --- Web framework ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0
aiofiles==23.2.1

# --- Data validation ---
pydantic==2.6.4
pydantic-settings==2.2.1

# --- Redis ---
redis==5.0.4

# --- PDF parsing ---
PyMuPDF==1.24.2

# --- LangGraph / LangChain ---
langchain==0.1.16
langgraph==0.0.30

# --- LLM providers ---
httpx==0.27.0
groq==0.5.0
google-generativeai==0.5.4

# --- STT ---
# Groq Whisper uses the groq package above
faster-whisper==1.0.3

# --- TTS ---
pyttsx3==2.90

# --- Web search ---
tavily-python==0.3.0

# --- Testing ---
pytest==8.1.1
pytest-asyncio==0.23.6
